docs: tell user if they use rhsa with no cves sooner (#829)
We already tell the user they can't use an RHSA with no cves in the
create-advisory task. However, the new set-severity-task runs before the
create-advisory task and will also fail if the user specifies RHSA with
no cves (it will say it couldn't find a severity for any cve). This
commit adds the same error message from create-advisory to
set-advisory-severity so it is more clear to the user.

Signed-off-by: Johnny Bieren <jbieren@redhat.com>