fix(KONFLUX-6580): error handling in get-image-architectures (#838)
All tasks that use get-image-architecture were updated to get a new
version that has `set -e` so that it fails if skopeo or oras fails.

It's better to fail rather than to get a bogus empty value as
a result of the script.

Also, modify create-pyxis-image to actually be able to fail if
`get-image-architecture` fails and add a test. (It turns out
failing something when doing `<<< $(command)` never fails the
script, even with `set -e`. The other places that call
`get-image-architecture` look ok.)

Signed-off-by: Martin Malina <mmalina@redhat.com>