fix(KONFLUX-7733): advisory_url to point to customer portal (#914)
The internal create-advisory-task task produces a result
called advisory_url. This is then used for the results file
(to be shown in Konflux UI) and it's also added as a comment
when closing Jiras in close-advisory-issues.

Until now, we would link to advisory file in the advisories
gitlab repo. That's not where we should be pointing users.
Instead, we will now link to customer portal, e.g.
https://access.redhat.com/errata/RHBA-2025:3373

Signed-off-by: Martin Malina <mmalina@redhat.com>