fix: bug in parallel processing of pushes (#931)
We used just one temp file for the source image docker config and
as a result, while a push for one component was running in
the background,
another component was already changing the contents
of that docker config temp file.

The fix is to use a new temp file for each component.
That way they won't interfere.

Signed-off-by: Martin Malina <mmalina@redhat.com>