feat(ISV-5733): release service uploads sboms to atlas v2 (#962)
The Atlas instance used by the release-service has been recently
upgraded to a version 2. With a new version the API contract is changing
and the upload code needs to be updated.

The new routes have been added to the collect-atlas-params and upload
code have been updated in upload-sbom-to-atlas. Changes have been
propagate to all relevant pipelines.

With this update Konflux no longer pushes SBOMs to Atlas v1 and switches
to v2 instead.

JIRA: ISV-5733

Signed-off-by: Ales Raszka <araszka@redhat.com>