feat(RELEASE-439): support multiple components in sign-index-image (#972)
The sign-index-image task now expect a new fbcResultsPath parameter
with a json file produced by add-fbc-contribution that contains
a list of references and digests to sign.

This way we will support multiple components to be signed.

Signing of all the target index images and manifest digests
is processed in parallel.

Signed-off-by: Martin Malina <mmalina@redhat.com>