Pubtool sign integration fixes (#1131)
* feat(CLOUDDST-25758): use pubtools-sign sign and upload images (#808)

internal request request-and-upload-signature was transfered to
use pubtools-sign and pubtools-pyxis to sign and upload signatures.
With that, internal requests are now called with multiple references and
digests. Batch size is limited by params string length which is set by
default to 4096. Retry mechanism in signing was also removed as it's
already included in pubtools-sign

Signed-off-by: Jindrich Luza <jluza@redhat.com>

* fix(CLOUDDST-25758): use pubtools-sign sign and upload images

References and digests are now correctly splitted into arrays

Signed-off-by: Jindrich Luza <jluza@redhat.com>

---------

Signed-off-by: Jindrich Luza <jluza@redhat.com>