feat(RELEASE-1675): auto generate README.md files for tasks/pipelines (#1157)
Add readme_generator.sh, which generates README.md files
for tasks and pipelines, as well as a check_readme.sh script and
workflow to check that the README.md files in each task and pipeline
match the output of readme-generator.sh. Also updated CONTRIBUTING.md
to mention this new requirement.

Add check in check_readme.sh to stop '.' from being
added to the end of task/pipeline param descriptions.

Signed-off-by: Paul Fitzgerald <pafitzge@redhat.com>