feat(RELEASE-1747): add public key for bugzilla issues (#1210)
Currently, only issues on the issues.redhat.com server have the public
key added. This commit modifies the embargo-check task to add it for all
issues. The goal of this is to ensure that bugzilla bzs have public
added. This will enable us to make the public key required for each
issue in the created advisory.yaml.

Signed-off-by: Johnny Bieren <jbieren@redhat.com>