fix: improve pipeline visualization (#1205)
Tasks in a pipeline are displayed according to the runAfter property.
When a task is configured to be able to run much earlier than the tasks
that depend on it, the edge visualization becomes more complex. When
these complexity-inducing tasks do not take much time to run, it doesn't
make as much sense to try to evaluate the tasks as soon as possible.

This commit adjusts the runAfter for a couple tasks which are generally
fast (seconds or less) in order to improve the overall visualization
complexity. The adjusted tasks include set-advisory-severity,
collect-cosign-params, and collect-atlas-params.

Signed-off-by: arewm <arewm@users.noreply.github.com>

rh-pre-commit.version: 2.3.2
rh-pre-commit.check-secrets: ENABLED