fix(KONFLUX-8821): Increase performance for large snapshots (#1218)
* fix(KONFLUX-8821): increase parallelization for pushing snapshot

In order to improve the default performance of pushing snapshots, we can
increase the concurrency to 20. This also requires the memory to be
increased.

Signed-off-by: arewm <arewm@users.noreply.github.com>

rh-pre-commit.version: 2.3.2
rh-pre-commit.check-secrets: ENABLED

* fix(KONFLUX-8821): increase the batch size for signing

When we are submitting larger Snapshots to be signed, we can improve
efficiency by increasing the batch sizes. This will enable more requests
to be processed at once. Since we are doing more with each request,
however, we also need to ensure that we allow for a longer timeout.

Signed-off-by: arewm <arewm@users.noreply.github.com>

rh-pre-commit.version: 2.3.2
rh-pre-commit.check-secrets: ENABLED

* fix(KONFLUX-8821): increase memory for create product sbom

Large snapshots may result in a quick spike in memory consumption. Even
though the memory didn't increase when viewing the pod metrics, the
containers were terminating with an OOMkill.

Signed-off-by: arewm <arewm@users.noreply.github.com>

rh-pre-commit.version: 2.3.2
rh-pre-commit.check-secrets: ENABLED

* fix(KONFLUX-8821): increase cosign concurrency and memory

This task can easily process additional concurrent signatures.
Increasing the concurrent signatures from the original default based on
performance testing with large snapshots. Even with the additioanl
concurrent signatures, the memory consumption remained below 2Gi, but I
am bumping it to allow for an additional factor of safety.

Signed-off-by: arewm <arewm@users.noreply.github.com>

rh-pre-commit.version: 2.3.2
rh-pre-commit.check-secrets: ENABLED