feat(ISV-6146): push regeneration data to S3 (#1238)
This implementation is a bit hacky and includes some code duplication,
but all related tasks will be removed within the next week or two. The
goal is to start collecting SBOM input data ASAP to support future
SBOM regeneration.

For context, we use the `release_ids` embedded in the SBOMs to match
them with their input data stored in the S3 bucket. This enables us to
regenerate SBOMs later if we discover issues in the generation
process, ensuring reliable data for incident response.