feat(RELEASE-1765): add collect-task-params managed task (#1245)
This commit adds a new task named collect-task-params. It is meant to
replace all collect-*-params managed tasks. At first, the primary use
case of this task will be to extract parameters from the data file for
the verify-conforma task. The pipelines will be updated to include this
task in a future commit.

Signed-off-by: Johnny Bieren <jbieren@redhat.com>