refactor(RELEASE-1777): remove collect-oot-sign-params task (#1276)
This commit removes the collect-oot-sign-params task in favor of the
collect-task-params task. This reduces the amount of code to maintain in
the repo and keeps the same functionality. The push-oot-kmods pipeline
was updated to use the collect-task-params task for these results. It
was also updated to include some results that were previously undefined
and used as parameters.

Signed-off-by: Johnny Bieren <jbieren@redhat.com>