fix(RELEASE-1785): user sparse checkout in filter task (#1287)
We started seeing issues when cloning the advisory repo
in the filter internal task. The task would get stuck
sometimes and eventually finish in ~40 minutes.
Based on my experiments, the memory should be
sufficient. So it's more likely about
disk IO.

Currently, cloning the full stage advisory repo
generates about 17k files. With this change, it's drastically
reduced depending on which tenant directory we need.
For dev-release-team-tenant, it went down to 312 files.

Disk size of the cloned repo went down from 138Mi to 6Mi.

Note that we have the same potential issue in
create-advisory-task internal task. That task uses a utils
function for the clone and it's also pushing a commit back
to gitlab, so it will be more complex and require more testing.
There will be a second commit/PR for that.

Also, I am aware it's not ideal this isn't really covered with
unit tests, but I'm not sure how I would do that - we probably
don't want to pull a real git repo in our tests.

Signed-off-by: Martin Malina <mmalina@redhat.com>