feat(RELEASE-1792): apply-mapping supports the repositories key
This commit updates the apply-mapping managed task to support the
mapping.component.repositories key. It behaves similar to the
mapping.component.repository key, but it is a list instead of a
singleton.

The commit maintains backwards compatibility with the repository key.
Also, if the repository key is passed, but not repositories,
repositories[0] will be set to what repository is.

Signed-off-by: Johnny Bieren <jbieren@redhat.com>