fix: gitlint check in merge queues take 3
Another attempt at fixing gitlint when it runs
as part of merge queues.

Previous attempts:
https://github.com/konflux-ci/release-service-catalog/pull/1341
https://github.com/konflux-ci/release-service-catalog/pull/1343

It still fails. The reason is that in the merge queue event
the base_ref will be something like `refs/heads/develoment`
and that will work for the fetch, but not as a branch name.
So we need to strip `refs/heads` from the base_ref.

Signed-off-by: Martin Malina <mmalina@redhat.com>