refactor(RELEASE-1797): create-pyxis-image uses repositories
This commit updates the create-pyxis-image managed task to read the
repositories key for each component instead of the repository key. This
is part of a larger effort for each component to specify multiple
repositories each instead of just one. Because apply-mapping emits both
repository and repositories now, the change does not have to be
backwards compatible.

Signed-off-by: Johnny Bieren <jbieren@redhat.com>