fix(RELEASE-1840): stop saving sboms in trusted artifact
Some users reported OOM when creating trusted artifact
in push-rpm-data-to-pyxis.

It turns out we were saving all the downloaded sboms
in the generated trusted artifact which is completely
necessary. In fact, this task wouldn't even need to generate
any TA at all - no other task will consume it. But it's convenient
for task tests, so I am not removing it completely. I am only
changing it so that the sboms are saved to another directory
inside the emptyDir volume.

Also, remove subdirectory from the task - it wasn't used for
anything anymore - it was probably an ommission from when
the task was transformed to support TA only.

Signed-off-by: Martin Malina <mmalina@redhat.com>