feat(ISV-6027): bump up Mobster version
A new version of Mobster tasks are available that contains update to how
purl in SBOMs are generated. The new version of purls contains 'arch'
identifier for oci images and index images.

This change is already available in the build part and this commits adds
a support also for a release service.

JIRA: ISV-6027

Signed-off-by: Ales Raszka <araszka@redhat.com>